% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_cytotrace_pipline}
\alias{run_cytotrace_pipline}
\title{run_cytotrace_pipline}
\usage{
run_cytotrace_pipline(
  rds,
  assay = NULL,
  slot = "counts",
  ncores = 1,
  chunk.size = 1000
)
}
\arguments{
\item{rds}{seurat object}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{ncores}{ncores Number of processors to parallelize computation. If \code{BPPARAM = NULL}, the function uses,\code{BiocParallel::MulticoreParam(workers=ncores)}}

\item{chunk.size}{How many cells to include in each sub-matrix}
}
\value{
seurat object
}
\description{
run_cytotrace_pipline
}
