% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_entropy}
\alias{run_entropy}
\title{run_entropy}
\usage{
run_entropy(
  rds,
  method,
  assay = NULL,
  slot = NULL,
  ncores = 1,
  species = NULL,
  chunk.size = 1000
)
}
\arguments{
\item{rds}{seurat object}

\item{method}{Method selection parameters for calculating cellular entropy，you can choose "cytotrace","slice","stemid","scent_sr" and "scent_ccat"}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{ncores}{ncores Number of processors to parallelize computation. If \code{BPPARAM = NULL}, the function uses,\code{BiocParallel::MulticoreParam(workers=ncores)}}

\item{species}{species choose,When the method is set to "scent_ccat" or "scent_sr", the species parameter can only be selected as "Human". When the method is set to "slice", the species parameter can be selected from Human, Mouse, and Rat. Other methods are not affected by the species parameter.}

\item{chunk.size}{How many cells to include in each sub-matrix}
}
\value{
seurat object
}
\description{
run_entropy
}
