% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_scent_ccat_pipline}
\alias{run_scent_ccat_pipline}
\title{run_scent_ccat_pipline}
\usage{
run_scent_ccat_pipline(rds, assay = NULL, slot = "counts", species = "Human")
}
\arguments{
\item{rds}{seurat object}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{species}{species choose,only choose Human}
}
\value{
seurat object
}
\description{
run_scent_ccat_pipline
}
