% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_scent_sr_pipline}
\alias{run_scent_sr_pipline}
\title{run_scent_sr_pipline}
\usage{
run_scent_sr_pipline(
  rds,
  assay = NULL,
  slot = "counts",
  species = "Human",
  ncores = 1
)
}
\arguments{
\item{rds}{seurat object}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{species}{species choose,only choose Human}

\item{ncores}{ncores Number of processors to parallelize computation. If \code{BPPARAM = NULL}, the function uses,\code{BiocParallel::MulticoreParam(workers=ncores)}}
}
\value{
seurat object
}
\description{
run_scent_sr_pipline
}
