% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_slice_pipline}
\alias{run_slice_pipline}
\title{run_slice_pipline}
\usage{
run_slice_pipline(
  rds,
  assay = NULL,
  slot = "data",
  species = "human",
  chunk.size = 1000
)
}
\arguments{
\item{rds}{seurat object}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}

\item{species}{species choose,only choose Human, Mouse,Rat}

\item{chunk.size}{How many cells to include in each sub-matrix}
}
\value{
seurat object
}
\description{
run_slice_pipline
}
