% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_stemid_pipline}
\alias{run_stemid_pipline}
\title{run_stemid_pipline}
\usage{
run_stemid_pipline(rds, assay = NULL, slot = "counts")
}
\arguments{
\item{rds}{seurat object}

\item{assay}{Pull out data from this assay of the Seurat object,(if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object,default is counts}
}
\value{
seurat object
}
\description{
run_stemid_pipline
}
